2-修正後
