--- 02/jtraining21/jtraining21.java
@@ -1,7 +1,0 @@
-package jtraining21;
-
-public class jtraining21 {		
-	 public static void main(String[] args) {	
-		System.out.println("Hello World!");
-	}	
-}
