--- 02/jtraining22/jtraining22.java
@@ -1,9 +1,0 @@
-package jtraining22;
-
-public class jtraining22 {		
-	 public static void main(String[] args) {
-		String x = "Hello";
-		String y = "World";
-		System.out.println(x +" "+ y);
-	}	
-}
