--- 02/jtraining23/jtraining23.java
@@ -1,13 +1,0 @@
-package jtraining23;
-
-public class jtraining23 {				
-		 public static void main(String[] args) {
-			int x = 4;
-			int y = 3;
-			
-			y = 4;
-			x = 3; 
-			System.out.println(x);
-			System.out.println(y);
-		}	
-}
