--- 02/jtraining24/jtraining24.java
@@ -1,12 +1,0 @@
-package jtraining24;
-
-public class jtraining24 {				
-		 public static void main(String[] args) {
-			int x = 7;
-
-			System.out.println(x*3);
-			System.out.println(x);
-			System.out.println(x/2);
-			System.out.println(x);
-		}	
-}
