+++ HelloWorld/src/jtraining21/Jtraining21.java
@@ -1,0 +1,14 @@
+package jtraining21;
+
+/**
+ * Hello World クラス
+ */
+public class Jtraining21 {		
+	/**
+	 * main　メソッド
+	 * param args　実行引数
+	 */
+	 public static void main(String[] args) {	
+		System.out.println("Hello World!");
+	}	
+}
