+++ HelloWorld/src/jtraining22/Jtraining22.java
@@ -1,0 +1,16 @@
+package jtraining22;
+
+/**
+ * Hello World クラス
+ */
+public class Jtraining22 {	
+	/**
+	 * main　メソッド
+	 * param args　実行引数
+	 */
+	 public static void main(String[] args) {
+		String x = "Hello";
+		String y = "World";
+		System.out.println(x + " " + y);
+	}	
+}
