+++ HelloWorld/src/jtraining24/Jtraining24.java
@@ -1,0 +1,19 @@
+package jtraining24;
+
+/**
+ * 掛け算割り算 クラス
+ */
+public class Jtraining24 {	
+	/**
+	 * main　メソッド
+	 * param args　実行引数
+	 */
+		 public static void main(String[] args) {
+			int x = 7;
+
+			System.out.println(x * 3);
+			System.out.println(x);
+			System.out.println(x / 2);
+			System.out.println(x);
+		}	
+}
